Add files via upload
The task at hand is to classify German Credit Data into good or bad credit risks. This classification problem is important in assessing creditworthiness and making informed lending decisions. The dataset contains various features such as savings and checking accounts, sex, housing, and purpose, among others. The goal is to build a model that accurately predicts the credit risk based on these features.